Double_dice_end/requirements.txt
asgiref==3.5.2
channels==3.0.5
Django==4.0.6
djangorestframework==3.14.0
